+++ hawk_sdk/__init__.py
@@ -1,1 +1,1 @@
-from hawk_sdk.futures import Futures
+from hawk_sdk.api.futures import Futures
