+++ hawk_sdk/api/futures/main.py
@@ -6,10 +6,10 @@ from typing import List
 
 from google.cloud import bigquery
 
-from hawk_sdk.common.constants import PROJECT_ID
-from hawk_sdk.common.data_object import DataObject
-from hawk_sdk.futures.repository import FuturesRepository
-from hawk_sdk.futures.service import FuturesService
+from hawk_sdk.core.common.constants import PROJECT_ID
+from hawk_sdk.core.common.data_object import DataObject
+from hawk_sdk.api.futures.repository import FuturesRepository
+from hawk_sdk.api.futures.service import FuturesService
 
 
 class Futures:
@@ -33,22 +33,4 @@ class Futures:
         return DataObject(
             name="futures_ohlcvo",
             data=self.service.get_ohlcvo(start_date, end_date, interval, hawk_ids)
-        )
-
-    def get_hgf_model_state(self, start_date: str, end_date: str, short_ema: int, long_ema: int) -> DataObject:
-        """Fetch Hawk Global Futures model state data for the given date range.
-
-        :param start_date: The start date for the data query (YYYY-MM-DD).
-        :param end_date: The end date for the data query (YYYY-MM-DD).
-        :param short_ema: The short exponential moving average period.
-        :param long_ema: The long exponential moving average period.
-        :return: A hawk DataObject containing the model state data.
-        :raises ValueError: If short EMA period is longer than long EMA period.
-        """
-        if short_ema > long_ema:
-            raise ValueError("Short EMA period cannot be longer than Long EMA period.")
-
-        return DataObject(
-            name="futures_hgf_model_state",
-            data=self.service.get_hgf_model_state(start_date, end_date, short_ema, long_ema)
         )
