+++ hawk_sdk/core/common/enum.py
@@ -1,0 +1,5 @@
+from hawk_sdk.core.common.base_enum import BaseEnum
+
+
+class Universe(BaseEnum):
+    HAWK_GLOBAL_FUTURES = "hawk_global_futures"
