--- hawk_sdk/futures/__init__.py
@@ -1,1 +1,0 @@
-from hawk_sdk.futures.main import Futures
